requirements.txt
pygsheets
python-dotenv
